requirements.txt
advertorch==0.1.4
cycler==0.10.0
imageio==2.4.1
kiwisolver==1.0.1
matplotlib==3.0.1
numpy==1.15.3
pandas==0.23.4
Pillow==5.3.0
progressbar==2.5
pyparsing==2.2.2
python-dateutil==2.7.4
pytz==2018.7
scikit-learn==0.20.3
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
torchvision==0.2.1
